algumas correções / main de testes para terminal

+++ headers/IStock.h
@@ -7,7 +7,7 @@ public:
 	/* obligatory getters */
 	virtual int getCap() = 0;
 	virtual std::string getName() = 0; // nome do produto armazenado
-	virtual float getPrice() = 0; // preco total do lote
+	virtual double getPrice() = 0; // preco total do lote
 	virtual int getQtt() = 0; // quantidade em estoque
 	/* obligatory setters */
 	virtual bool setCap(int) = 0;
